% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IComp.R
\name{IComp}
\alias{IComp}
\title{Computes the Importance Index of a transcription factor in a set of association rules}
\usage{
IComp(TFi, rules_TF, rules_noTF, figures)
}
\arguments{
\item{TFi}{string or string vector: the transcription factor (or combination of transcription factors) whose importance distribution is evaluated}

\item{rules_TF}{a set of rules in which \code{TFi} is present}

\item{rules_noTF}{a set of rules obtained from rules_TF removing transcription factor (or combination of transcription factors) in TFi (obrained with the function \code{\link{rulesNTF}})}

\item{figures}{logical if \code{figures = TRUE}, graphics with support, lift and confidence distributions of the rules in \code{rules_TF} and \code{rules_noTF} are returned}
}
\value{
A list of two elements: the \code{imp} element is a vector of integers with the importances of TFi in each rule in \code{rules_TF}; the \code{delta} element of the list is a list with variations of standardized distributions of the three measures of support, confidence and lift. This output is used in the function \code{\link{IPCA}} for the Principal Component Analysis of such distributions.
}
\description{
Given an association rule and a transcription factor \code{TFi}, it is evaluated the contribution of \code{TFi} in the rule for the prediction of the presence of the item in the right-hand-side of the rule.
Since this contribution is evaluated based on the variations of support, confidence and lift of the rule, the user can visualize such variations setting the parameter \code{figures = TRUE}.
}
\examples{
# Load the data:
data("data_man")

# The Importance Indexes of FOSL2=1 in the set of rules r_FOSL2 are given by:
IComp("FOSL2=1", r_FOSL2, r_noFOSL2, figures=TRUE)
}

