% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMP_Z.R
\docType{data}
\name{IMP_Z}
\alias{IMP_Z}
\title{Contains the Importance Index associated with each co-regulator
which is present in at least one association rule.}
\format{An object of class \code{"list"}}
\usage{
data("data_man")
}
\description{
IMP_Z is a list of 12 elements, containig the Importance Indexes for all
transcription factors (\code{p}) in the set of relevant association rules
extracted. It is included in the \code{data_man} data collection.
}
\examples{
# IMP_Z is found in the data_man collection of datasets:
data("data_man")
head(IMP_Z[[1]])
}
\keyword{datasets}
