% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{items}
\alias{items}
\title{Extracts items in the left-hand-side of an association rule.}
\usage{
items(itemset)
}
\arguments{
\item{itemset}{object of class rules.}
}
\value{
A string vector with the items in the left-hand-side
of the rule in itemset.
}
\description{
The function is used in the elaboration of the left-hand-side of
association rules to search for the items.
}
\examples{
items('{TAF1=1,EP300=1,MAX=1}')
# the output is: 'TAF1=1','EP300=1','MAX=1'
}
