% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemset.R
\name{itemset}
\alias{itemset}
\title{Builds the itemset in the left-hand-side of an association rule.}
\usage{
itemset(items)
}
\arguments{
\item{items}{a string vector.}
}
\value{
Itemset of the left-hand-side of a searched rule, with the items
in \code{items}.
}
\description{
The function is used in the construction of the left-hand-side of association
rules to search for a rule of interest.
}
\examples{
itemset(c('TAF1=1','EP300=1'))
# the output is: 'TAF1=1,EP300=1'
}
