% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\docType{data}
\name{p}
\alias{p}
\title{Contains co-regulators found in at least one association rule.}
\format{An object of class \code{"character"}}
\usage{
data("data_man")
}
\description{
p is a character vector that contains the transcription factors
present in at least one association rule. It is included in the 
\code{data_man} data collection.
}
\examples{
# p is found in the data_man collection of datasets:
data("data_man")
p
}
\keyword{datasets}
