% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presAbs.R
\name{presAbs}
\alias{presAbs}
\title{Splits a set of transcription factors in 'present' or 'absent' transcription factors in a set of rules.}
\usage{
presAbs(TFs, rules, type)
}
\arguments{
\item{TFs}{a string vector: set of transcription factors to search in the set of rules.}

\item{rules}{data.frame: rules and their quality measures (support, confidence and lift).}

\item{type}{logical parameter: if \code{type = TRUE} only rules with present transcription factors in the left-hand-side are considered.}
}
\value{
A list of two string vectors: the list \code{pres} contains all the transcription factors in \code{TFs} that are present in \code{rules}, and the list \code{abs} contains all the transcription factors in \code{TFs} that are absent in \code{rules}.
}
\description{
The function is used to find candidate co-regulator transcription factors with a given transcription factor, by finding the transcription factors that contribute to the prediction of its presence in a set of association rules.
}

