% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulesNTF.R
\name{rulesNTF}
\alias{rulesNTF}
\title{Removes a transcription factor (or a combination of transcription factors) from the left-hand-side of a set of rules.}
\usage{
rulesNTF(TFi, sub_rules, all_rules)
}
\arguments{
\item{TFi}{a string, or a string vector: transcription factor (or combination of transcription factors) to remove from the set of rules.}

\item{sub_rules}{a data.frame with a subset of rules containing \code{TFi}, and their quality measures of support, confidence and lift (i.e., rules from which the user wants to remove \code{TFi}).}

\item{all_rules}{a data.frame with a set of rules and their quality measures of support, confidence and lift.}
}
\value{
a data.frame with all the rules in the set \code{rules} without the transcription factor (or combination of transcription factors) in \code{TFi}, and their quality measures of support, confidence and lift.
}
\description{
The function removes a given transcription factor \code{TFi} (or a combination of transcription factors) chosen by the user, from the subset of relevant association rules extracted with the function \code{\link{rulesTF}}. Then it searches for the obtained rules and their quality measures of support, confidence and lift
in the set of most relevant associations extracted with the function \code{\link{rulesGen}}. If a rule is not found in such set, its quality measures are set to 0.
}
\examples{
# Load the data:
data("data_man")

r_noFOSL2 <- rulesNTF("FOSL2=1", r_FOSL2, r_TEAD4)

}

