% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_rule.R
\name{search_rule}
\alias{search_rule}
\title{Finds an association rule in a dataset.}
\usage{
search_rule(data, LHS, RHS)
}
\arguments{
\item{data}{a binary matrix or data.frame in which the rule is searched}

\item{LHS}{a string with the left-hand-side of the searched rule}

\item{RHS}{a string with the right-hand-side of the searched rule}
}
\value{
A vector with five elements: the left-hand-side of the searched rule,
the right-hand-side of the searched rule, the support, confidence and lift
of the found rule.
}
\description{
The function looks for an association rule in a dataset \code{data}
with a fixed left-hand-side (\code{LHS}) and a fixed right-hand-side
(\code{RHS}). The function is used in the \code{rulesTF0}.
}
